crawler/requirements.txt
requests==2.31.0
selenium==4.15.2
webdriver_manager==4.0.1
pandas==2.1.3
python-dotenv==1.0.0
fake-useragent==1.4.0 